+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.1...HEAD) - TBD
 
+### :bug: Fixed
+
+- Prevent response cookies from leaking between generated requests when Schemathesis reuses an engine-managed `requests.Session`.
+
 ## [4.21.1](https://github.com/schemathesis/schemathesis/compare/v4.21.0...v4.21.1) - 2026-06-06
 
 ### :bug: Fixed